make sure we keep citation recording

+++ data_to_paper/data_to_paper/base_steps/base_steps_runner.py
@@ -109,7 +109,8 @@ class BaseStepsRunner(ProductsHandler):
             # delete all the files except the mock_openai file:
             for file in glob.glob(str(self.output_directory / '*')):
                 if file != str(self.output_directory / self.OPENAI_RESPONSES_FILENAME) \
-                        and file != str(self.output_directory / self.CROSSREF_RESPONSES_FILENAME):
+                        and file != str(self.output_directory / self.CROSSREF_RESPONSES_FILENAME) \
+                        and file != str(self.output_directory / self.SEMANTIC_SCHOLAR_RESPONSES_FILENAME):
                     os.remove(file)
         else:
             os.makedirs(self.output_directory)
